Move QgsCompositionChecker to core, speeds up compilation with tests enabled
