+++ src/core/qgsmultirenderchecker.cpp
@@ -14,7 +14,7 @@
  ***************************************************************************/
 
 #include "qgsmultirenderchecker.h"
-
+#include "qgscomposition.h"
 #include <QDebug>
 
 QgsMultiRenderChecker::QgsMultiRenderChecker()
@@ -102,3 +102,84 @@ QString QgsMultiRenderChecker::controlImagePath() const
                               QDir::separator() + mControlPathPrefix + QDir::separator() + mControlName + QDir::separator();
   return myControlImageDir;
 }
+
+#ifdef ENABLE_TESTS
+
+//
+// QgsCompositionChecker
+//
+
+///@cond PRIVATE
+
+QgsCompositionChecker::QgsCompositionChecker( const QString& testName, QgsComposition* composition )
+    : QgsMultiRenderChecker()
+    , mTestName( testName )
+    , mComposition( composition )
+    , mSize( 1122, 794 )
+    , mDotsPerMeter( 96 / 25.4 * 1000 )
+{
+  // The composer has some slight render inconsistencies on the whole image sometimes
+  setColorTolerance( 5 );
+}
+
+QgsCompositionChecker::QgsCompositionChecker()
+    : mComposition( nullptr )
+    , mDotsPerMeter( 96 / 25.4 * 1000 )
+{
+}
+
+QgsCompositionChecker::~QgsCompositionChecker()
+{
+}
+
+bool QgsCompositionChecker::testComposition( QString &theReport, int page, int pixelDiff )
+{
+  if ( !mComposition )
+  {
+    return false;
+  }
+
+  setControlName( "expected_" + mTestName );
+
+#if 0
+  //fake mode to generate expected image
+  //assume 96 dpi and size of the control image 1122 * 794
+  QImage newImage( QSize( 1122, 794 ), QImage::Format_RGB32 );
+  mComposition->setPlotStyle( QgsComposition::Print );
+  newImage.setDotsPerMeterX( 96 / 25.4 * 1000 );
+  newImage.setDotsPerMeterY( 96 / 25.4 * 1000 );
+  drawBackground( &newImage );
+  QPainter expectedPainter( &newImage );
+  //QRectF sourceArea( 0, 0, mComposition->paperWidth(), mComposition->paperHeight() );
+  //QRectF targetArea( 0, 0, 3507, 2480 );
+  mComposition->renderPage( &expectedPainter, page );
+  expectedPainter.end();
+  newImage.save( mExpectedImageFile, "PNG" );
+  return true;
+#endif //0
+
+  QImage outputImage( mSize, QImage::Format_RGB32 );
+
+  mComposition->setPlotStyle( QgsComposition::Print );
+  outputImage.setDotsPerMeterX( mDotsPerMeter );
+  outputImage.setDotsPerMeterY( mDotsPerMeter );
+  drawBackground( &outputImage );
+  QPainter p( &outputImage );
+  mComposition->renderPage( &p, page );
+  p.end();
+
+  QString renderedFilePath = QDir::tempPath() + '/' + QFileInfo( mTestName ).baseName() + "_rendered.png";
+  outputImage.save( renderedFilePath, "PNG" );
+
+  setRenderedImage( renderedFilePath );
+
+  bool testResult = runTest( mTestName, pixelDiff );
+
+  theReport += report();
+
+  return testResult;
+}
+
+///@endcond
+
+#endif
