+++ tests/src/core/testqgsatlascomposition.cpp
@@ -17,7 +17,7 @@
 
 #include "qgsapplication.h"
 #include "qgscomposition.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgscomposermap.h"
 #include "qgscomposermapoverview.h"
 #include "qgsatlascomposition.h"
