+++ tests/src/core/testqgscomposereffects.cpp
@@ -17,7 +17,7 @@
 
 #include "qgsapplication.h"
 #include "qgscomposition.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgscomposershape.h"
 #include "qgsmaprenderer.h"
 #include <QObject>
