+++ tests/src/core/testqgscomposergroup.cpp
@@ -18,7 +18,7 @@
 #include "qgscomposeritemgroup.h"
 #include "qgscomposerlabel.h"
 #include "qgscomposition.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgsapplication.h"
 
 #include <QObject>
