+++ tests/src/core/testqgscomposermultiframe.cpp
@@ -19,7 +19,7 @@
 #include "qgscomposerframe.h"
 #include "qgscomposerlabel.h"
 #include "qgscomposition.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgsapplication.h"
 
 #include <QObject>
