+++ tests/src/core/testqgscomposerpaper.cpp
@@ -17,7 +17,7 @@
 
 #include "qgsapplication.h"
 #include "qgscomposition.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgscomposershape.h"
 #include "qgssymbolv2.h"
 #include "qgssinglesymbolrendererv2.h"
