+++ tests/src/core/testqgscomposerpicture.cpp
@@ -17,7 +17,7 @@
 
 #include "qgsapplication.h"
 #include "qgscomposition.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgscomposerpicture.h"
 #include <QObject>
 #include <QtTest/QtTest>
