+++ tests/src/core/testqgscomposerrotation.cpp
@@ -17,7 +17,7 @@
 
 #include "qgsapplication.h"
 #include "qgscomposition.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgscomposershape.h"
 #include "qgscomposermap.h"
 #include "qgscomposerlabel.h"
