+++ tests/src/core/testqgscomposerscalebar.cpp
@@ -17,7 +17,7 @@
 
 #include "qgsapplication.h"
 #include "qgscomposition.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgscomposermap.h"
 #include "qgscomposerscalebar.h"
 #include "qgsmaplayerregistry.h"
