+++ tests/src/core/testqgscomposertablev2.cpp
@@ -25,7 +25,7 @@
 #include "qgsvectorlayer.h"
 #include "qgsvectordataprovider.h"
 #include "qgsfeature.h"
-#include "qgscompositionchecker.h"
+#include "qgsmultirenderchecker.h"
 #include "qgsfontutils.h"
 #include "qgsmaplayerregistry.h"
 #include "qgsproject.h"
